Rework cell capacity metrics and alerts
We recently introduced isolation segments in to our production environments.
Isolation segments are collections of diego cells that are reserved for the
exclusive used of the Cloud Foundry orgs that are granted access to them. Aside
from this and their instance type, they exhibit identical characteristics.

For a long time we’ve a chart on our monitoring dashboard that shows us the
number of cells we have deployed, versus the number of cells we think we need
to meet our cell capacity goals, as per ADR021 [1]. With the introduction of
isolation segments, this chart has been misreporting the number of deployed
cells (and thus the required number), because the metrics haven’t distinguished
between different groups of cells.

On top of that, the metrics used to calculate the used capacity and required
capacity was confusing and complicated.

This commit re-works those metrics, to make them easier to maintain, and to
have them differentiate between the cell groups. It adds 4 new metrics:

1.`rep_memory_used_pct:avg5m` the average amount of memory used in a 5m window
over all cells in a group. This is the inverse of
`rep_memory_capacity_pct:avg5m`

2. `diego_cells_deployed` the number of cells currently deployed in each group

3. `diego_cell_capacity_used_pct:avg5m` what percentage of the deployed cells’
capacity is being used. This is calculated by multiplying the average memory
usage percentage by the number of cells. For example, if you have 10 cells
deployed, with an average memory utilisation of 35%, then you have 10 * 35% =
3.5 cells utilised.

4. `diego_cells_required` the number of cells required to meet our capacity
goals as per [1]. This is calculated at 1.5 times the number of cells being
utilised. Our goal is to have 1/3rd spare capacity at all times, so to work out
how many cells we need, we must treat the number of utilised cells as 2/3rd of
our capacity. Multiplying by 1.5 achieves that.

The commit also adds some useful tests of expressions, and ensures the cell
memory capacity alarm alerts us as to which group of cells is running under
capacity.

[1] https://team-manual.cloud.service.gov.uk/architecture_decision_records/ADR021-cell-capacity-assignment-2/